redo codegen to be next to operation files and simpler
