+++ apps/web/src/api/server/contentful/fetchCurrentLocation.ts
@@ -1,8 +1,8 @@
 import { gql } from 'graphql-request';
 import { isNotNullish } from 'shared-core/helpers/typeguards';
-import type { MyLocationQuery } from 'api/types/generated/fetchCurrentLocation.generated';
 import type { MapLocation } from 'api/types/MapLocation';
 import { contentfulClient } from '../networkClients/contentfulClient';
+import type { MyLocationQuery } from './fetchCurrentLocation.generated';
 
 /**
  * Grabs the home location using a known id for it
