+++ apps/web/src/api/server/contentful/fetchFooterLinks.generated.ts
@@ -1,4 +1,4 @@
-import type * as Types from './contentfulApi.generated';
+import type * as Types from './api.generated';
 
 export type FooterQueryVariables = Types.Exact<{ [key: string]: never }>;
