+++ apps/web/src/api/server/contentful/fetchProjects.generated.ts
@@ -1,4 +1,4 @@
-import type * as Types from './contentfulApi.generated';
+import type * as Types from './api.generated';
 
 export type ProjectsQueryVariables = Types.Exact<{ [key: string]: never }>;
