+++ apps/web/src/components/RichText.tsx
@@ -7,7 +7,7 @@ import { isNotNullish } from 'shared-core/helpers/typeguards';
 import type { SxProps } from 'ui/theme';
 import { PROJECT_MAX_IMAGE_DIMENSION } from 'ui/helpers/imageSizes';
 import { isLink, isProject } from 'api/parsers';
-import type { Asset, Entry, TextBlockContent } from 'api/types/generated/contentfulApi.generated';
+import type { Asset, Entry, TextBlockContent } from 'api/server/contentful/api.generated';
 import { ProjectCard } from 'components/homepage/ProjectCard';
 import { Image } from './Image';
 import { Link } from './Link';
