+++ packages/codegen/package.json
@@ -2,6 +2,7 @@
   "name": "codegen",
   "private": true,
   "scripts": {
+    "clean": "find ../../ -type f -name '*.generated.ts' -delete",
     "codegen": "graphql-codegen --config config.ts",
     "format": "prettier --write .",
     "lint": "eslint . --fix",
