+++ packages/eslint-config-dg/library.js
@@ -44,7 +44,6 @@ module.exports = {
     '.vercel/',
     '.turbo/',
     '**/*.generated.*',
-    '**/generated/*',
     'migrations/',
   ],
   rules,
